Add an UPDATE_FEATURE_METADATA event.
This conveys feature metadata from the user agent to other.js. It'll be useful
for writing a message action that knows how to install arbitrary features.

Ref https://github.com/other-xyz/other-chat-web/issues/289
