+++ lib/Events.js
@@ -12,6 +12,7 @@ export const SET_MESSAGE_ACTIONS = 'SET_MESSAGE_ACTIONS'
 export const SET_SELECTED_MESSAGES = 'SET_SELECTED_MESSAGES'
 export const SET_STAGED_MESSAGE = 'SET_STAGED_MESSAGE'
 export const UPDATE_ENTITIES = 'UPDATE_ENTITIES'
+export const UPDATE_FEATURE_METADATA = 'UPDATE_FEATURE_METADATA'
 export const UPDATE_MESSAGES = 'UPDATE_MESSAGES'
 export const UPDATE_STAGED_MESSAGE = 'UPDATE_STAGED_MESSAGE'
