+++ lib/Feature.js
@@ -1,6 +1,7 @@
 import * as Events from './Events'
 import Chatternet from './Chatternet'
 import CommandListener from './CommandListener'
+import FeatureMetadata from './FeatureMetadata'
 import MentionListener from './MentionListener'
 import TokenListener from './TokenListener'
 import WordListener from './WordListener'
@@ -16,22 +17,19 @@ import {userAgent} from './userAgent'
  * In the later case, the feature is available for any user visiting the
  * channel.
  */
-class Feature {
+class Feature extends FeatureMetadata {
   /**
-   * @param {string} name - User facing name.
    * @param {string} description - User facing, one-line description.
-   * @param {string} version - {@link http://semver.org/|Semantic Version} used
-   *     for determining compatibility with clients and other features.
    * @param {string} identity - Feature developer's Chatternet identity under
    *     which all Chatternet operations are performed.
-   * @param {Command[]} commands - Commands to register.
+   * @param {Listener[]} listeners - Listeners to register.
+   * @param {string} name - User facing name.
+   * @param {string} version - {@link http://semver.org/|Semantic Version} used
+   *     for determining compatibility with clients and other features.
    */
-  constructor({name, description, version, identity, listeners = []}) {
+  constructor({description, identity, listeners = [], name, version}) {
+    super({description, identity, name, version})
     // TODO: Validation
-    this.name = name
-    this.description = description
-    this.version = version
-    this.identity = identity
     this._chatternet = new Chatternet({identity})
     this._listeners = []
     this._isListening = false
@@ -92,6 +90,7 @@ class Feature {
   _startListeningIfNeeded() {
     if (this._isListening) return
     this._isListening = true
+    this.isUserAgentRequired = true
 
     userAgent.on(Events.ACTIVATE_CHAT_COMPLETE_RESULT, (event) => {
       const {action, message, result, tag} = event
@@ -140,6 +139,7 @@ class Feature {
   _startProvidingMessageActionsIfNeeded() {
     if (this._isProvidingMessageActions) return
     this._isProvidingMessageActions = true
+    this.isUserAgentRequired = true
 
     userAgent.on(Events.SET_SELECTED_MESSAGES, (event) => {
       const {messages, tag} = event
