+++ lib/embedder/FeatureHost.js
@@ -1,5 +1,6 @@
 import * as events from '../Events'
 import ChatternetRestClient from './ChatternetRestClient'
+import FeatureMetadata from '../FeatureMetadata'
 
 /**
  * A host for a single feature.
@@ -7,48 +8,14 @@ import ChatternetRestClient from './ChatternetRestClient'
  * Useful for embedders to discover metadata, enforce
  * permissions and handle certain events.
  */
-export default class FeatureHost {
+export default class FeatureHost extends FeatureMetadata {
   /**
    * @param {!Object} module The module exported by the feature.
    * @param {!string} url The feature's URL.
    */
   constructor(module, url) {
-    this.module = module
-    this.url = url
+    super(Object.assign({url}, module))
     this._client = null
-  }
-
-  get name() {
-    return this.module.name
-  }
-
-  get version() {
-    return this.module.version
-  }
-
-  get identity() {
-    return this.module.identity
-  }
-
-  get isCloudEmbeddable() {
-    return !this.module._userAgent
-  }
-
-  get isUserAgentEmbeddable() {
-    return !this.module.identity
-  }
-
-  get displayUrl() {
-    return this.url
-        .replace('https://apps.other.chat/', '')
-        .replace('https://gist.githubusercontent.com/', '')
-  }
-
-  get icon() {
-    if (this.url.startsWith('https://gist.githubusercontent.com/')) {
-      return 'https://assets-cdn.github.com/favicon.ico'
-    }
-    return 'https://other.chat/images/favicon.png'
   }
 
   get client() {
