+++ lib/userAgent.js
@@ -14,6 +14,7 @@ import EventEmitter from 'events'
  * @listens UserAgent#SET_ACTIVE_IDENTITY
  * @listens UserAgent#SET_SELECTED_MESSAGES
  * @listens UserAgent#SET_STAGED_MESSAGE
+ * @listens UserAgent#UPDATE_FEATURE_METADATA
  * @inheritdoc
  */
 class UserAgent extends EventEmitter {
@@ -116,11 +117,22 @@ class UserAgent extends EventEmitter {
    *     make to the staged message, i.e. omitted fields remain unchanged.
    */
 
+  /**
+   * Event conveying that known feature metadata has been updated.
+   * @event UserAgent#UPDATE_FEATURE_METADATA
+   * @mixes event:TAGGED_MESSAGE
+   * @type {!Object} metadata - Sparse update dict of URL to metadata.
+   *     Null values represent a delete.
+   */
   constructor() {
     super()
     this._identity = null
+    this.featureMetadata = {}
     this.on(Events.SET_ACTIVE_IDENTITY, (event) => {
       this._identity = event.identity
+    })
+    this.on(Events.UPDATE_FEATURE_METADATA, (event) => {
+      this.featureMetadata = Object.assign(this.featureMetadata, event.metadata)
     })
   }
