+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "other",
-  "version": "3.8.2",
+  "version": "3.9.0",
   "description": "The Chatternet feature platform",
   "author": "Tony Gentilcore <tony@other.xyz>",
   "repository": "other-xyz/other.js",
